Update upgrade_test arch to be used in other pkgs.
