+++ pkg/pgmodel/migrate.go
@@ -14,6 +14,7 @@ import (
 	"path/filepath"
 	"regexp"
 	"sort"
+	"strconv"
 	"strings"
 	"sync"
 
@@ -367,6 +368,7 @@ func replaceSchemaNames(r io.ReadCloser) (string, error) {
 	}
 	s := buf.String()
 	s = strings.ReplaceAll(s, "SCHEMA_CATALOG", schema.Catalog)
+	s = strings.ReplaceAll(s, "SCHEMA_LOCK_ID", strconv.Itoa(schema.LockID))
 	s = strings.ReplaceAll(s, "SCHEMA_EXT", schema.Ext)
 	s = strings.ReplaceAll(s, "SCHEMA_PROM", schema.Prom)
 	s = strings.ReplaceAll(s, "SCHEMA_SERIES", schema.SeriesView)
