+++ pkg/tests/upgrade_tests/upgrade_test.go
@@ -15,7 +15,6 @@ import (
 	"os"
 	"os/exec"
 	"reflect"
-	"regexp"
 	"testing"
 
 	"github.com/blang/semver/v4"
@@ -78,7 +77,7 @@ func TestUpgradeFromPrev(t *testing.T) {
 	pristineDbInfo := getPristineDbInfo(t, false, baseExtensionState)
 
 	if !reflect.DeepEqual(pristineDbInfo, upgradedDbInfo) {
-		printDbInfoDifferences(t, pristineDbInfo, upgradedDbInfo)
+		PrintDbSnapshotDifferences(t, pristineDbInfo, upgradedDbInfo)
 	}
 }
 
@@ -89,7 +88,7 @@ func TestUpgradeFromPrevMultinode(t *testing.T) {
 	pristineDbInfo := getPristineDbInfo(t, false, extState)
 
 	if !reflect.DeepEqual(pristineDbInfo, upgradedDbInfo) {
-		printDbInfoDifferences(t, pristineDbInfo, upgradedDbInfo)
+		PrintDbSnapshotDifferences(t, pristineDbInfo, upgradedDbInfo)
 	}
 }
 
@@ -100,11 +99,11 @@ func TestUpgradeFromPrevNoData(t *testing.T) {
 	pristineDbInfo := getPristineDbInfo(t, true, baseExtensionState)
 
 	if !reflect.DeepEqual(pristineDbInfo, upgradedDbInfo) {
-		printDbInfoDifferences(t, pristineDbInfo, upgradedDbInfo)
+		PrintDbSnapshotDifferences(t, pristineDbInfo, upgradedDbInfo)
 	}
 }
 
-func getUpgradedDbInfo(t *testing.T, noData bool, extensionState testhelpers.ExtensionState) (upgradedDbInfo dbInfo) {
+func getUpgradedDbInfo(t *testing.T, noData bool, extensionState testhelpers.ExtensionState) (upgradedDbInfo dbSnapshot) {
 	// we test that upgrading from the previous version gives the correct output
 	// by induction, this property should hold true for any chain of versions
 	prevVersion := semver.MustParse(version.EarliestUpgradeTestVersion)
@@ -148,12 +147,12 @@ func getUpgradedDbInfo(t *testing.T, noData bool, extensionState testhelpers.Ext
 				ingestor.Close()
 
 			}
-			upgradedDbInfo = getDbInfo(t, dbContainer, dbTmpDir, db)
+			upgradedDbInfo = SnapshotDB(t, dbContainer, *testDatabase, dbTmpDir, db, true)
 		})
 	return
 }
 
-func getPristineDbInfo(t *testing.T, noData bool, extensionState testhelpers.ExtensionState) (pristineDbInfo dbInfo) {
+func getPristineDbInfo(t *testing.T, noData bool, extensionState testhelpers.ExtensionState) (pristineDbInfo dbSnapshot) {
 	withNewDBAtCurrentVersion(t, *testDatabase, extensionState,
 		/* preRestart */
 		func(container testcontainers.Container, _ string, db *pgxpool.Pool, tmpDir string) {
@@ -179,7 +178,7 @@ func getPristineDbInfo(t *testing.T, noData bool, extensionState testhelpers.Ext
 
 				doIngest(t, ingestor, postUpgradeData1, postUpgradeData2)
 			}
-			pristineDbInfo = getDbInfo(t, container, tmpDir, db)
+			pristineDbInfo = SnapshotDB(t, container, *testDatabase, tmpDir, db, true)
 		})
 	return
 }
@@ -235,54 +234,6 @@ var (
 		},
 	}
 )
-
-func printDbInfoDifferences(t *testing.T, pristineDbInfo dbInfo, upgradedDbInfo dbInfo) {
-	t.Errorf("upgrade differences")
-	if !reflect.DeepEqual(upgradedDbInfo.schemaNames, pristineDbInfo.schemaNames) {
-		t.Logf("different schemas\nexpected:\n\t%v\ngot:\n\t%v", pristineDbInfo.schemaNames, upgradedDbInfo.schemaNames)
-	}
-	if !reflect.DeepEqual(upgradedDbInfo.extensions, pristineDbInfo.extensions) {
-		t.Logf("different extensions\nexpected:\n\t%v\ngot:\n\t%v", pristineDbInfo.extensions, upgradedDbInfo.extensions)
-	}
-	pristineSchemas := make(map[string]schemaInfo)
-	for _, schema := range pristineDbInfo.schemas {
-		pristineSchemas[schema.name] = schema
-	}
-	for _, schema := range upgradedDbInfo.schemas {
-		expected, ok := pristineSchemas[schema.name]
-		if !ok {
-			t.Logf("extra schema %s", schema.name)
-			continue
-		}
-		tablesDiff := schema.tables != expected.tables
-		functionsDiff := schema.functions != expected.functions
-		privilegesDiff := schema.privileges != expected.privileges
-		indicesDiff := schema.indices != expected.indices
-		triggersDiff := schema.triggers != expected.triggers
-		dataDiff := !reflect.DeepEqual(schema.data, expected.data)
-		if tablesDiff || functionsDiff || privilegesDiff || indicesDiff || triggersDiff || dataDiff {
-			t.Logf("differences in schema: %s", schema.name)
-		}
-		if tablesDiff {
-			t.Logf("tables\nexpected:\n\t%s\ngot:\n\t%s", expected.tables, schema.tables)
-		}
-		if functionsDiff {
-			t.Logf("functions\nexpected:\n\t%s\ngot:\n\t%s", expected.functions, schema.functions)
-		}
-		if privilegesDiff {
-			t.Logf("privileges\nexpected:\n\t%s\ngot:\n\t%s", expected.privileges, schema.privileges)
-		}
-		if indicesDiff {
-			t.Logf("indices\nexpected:\n\t%s\ngot:\n\t%s", expected.indices, schema.indices)
-		}
-		if triggersDiff {
-			t.Logf("triggers\nexpected:\n\t%s\ngot:\n\t%s", expected.triggers, schema.triggers)
-		}
-		if dataDiff {
-			t.Logf("data\nexpected:\n\t%+v\ngot:\n\t%+v", expected.data, schema.data)
-		}
-	}
-}
 
 func addNode2(t testing.TB, DBName string) {
 	db, err := pgx.Connect(context.Background(), testhelpers.PgConnectURL(DBName, testhelpers.Superuser))
@@ -519,166 +470,6 @@ func doIngest(t *testing.T, ingestor *ingestor.DBIngestor, data ...[]prompb.Time
 		}
 		_ = ingestor.CompleteMetricCreation()
 	}
-}
-
-var schemas []string = []string{
-	"_prom_catalog",
-	"_prom_ext",
-	"_timescaledb_cache",
-	"_timescaledb_catalog",
-	"_timescaledb_config",
-	"_timescaledb_internal",
-	"information_schema",
-	"pg_catalog",
-	"pg_temp_1",
-	"pg_toast",
-	"pg_toast_temp_1",
-	"prom_api",
-	"prom_data",
-	"prom_data_series",
-	"prom_info",
-	"prom_metric",
-	"prom_series",
-	"public",
-	"timescaledb_information",
-}
-
-var ourSchemas []string = []string{
-	"public",
-	"_prom_catalog",
-	"_prom_ext",
-	"prom_api",
-	"prom_data",
-	"prom_data_series",
-	"prom_info",
-	"prom_metric",
-	"prom_series",
-}
-
-type dbInfo struct {
-	schemaNames []string
-	schemas     []schemaInfo
-	extensions  string // \dx
-}
-
-type schemaInfo struct {
-	name       string
-	tables     string // \d+
-	functions  string // \df+
-	privileges string // \dp+
-	indices    string // \di
-	triggers   string // \dy
-	data       []tableInfo
-}
-
-type tableInfo struct {
-	name   string
-	values []string
-}
-
-var replaceChildren = regexp.MustCompile("timescaledb_internal\\._hyper_.*\n")
-
-func getDbInfo(t *testing.T, container testcontainers.Container, outputDir string, db *pgxpool.Pool) (info dbInfo) {
-
-	info.schemaNames = getSchemas(t, db)
-	if !reflect.DeepEqual(info.schemaNames, schemas) {
-		t.Errorf(
-			"unexpected schemas.\nexpected\n\t%v\ngot\n\t%v",
-			info.schemas,
-			schemas,
-		)
-	}
-
-	info.extensions = getPsqlInfo(t, container, outputDir, "\\dx")
-	info.schemas = make([]schemaInfo, len(ourSchemas))
-	for i, schema := range ourSchemas {
-		info := &info.schemas[i]
-		info.name = schema
-		info.tables = getPsqlInfo(t, container, outputDir, "\\d+ "+schema+".*")
-		info.tables = replaceChildren.ReplaceAllLiteralString(info.tables, "timescaledb_internal._hyper_*\n")
-		info.functions = getPsqlInfo(t, container, outputDir, "\\df+ "+schema+".*")
-		info.privileges = getPsqlInfo(t, container, outputDir, "\\dp "+schema+".*")
-		// not using \di+ since the sizes are too noisy, and the descriptions
-		// will be in tables anyway
-		info.indices = getPsqlInfo(t, container, outputDir, "\\di "+schema+".*")
-		info.triggers = getPsqlInfo(t, container, outputDir, "\\dy "+schema+".*")
-		info.data = getTableInfosForSchema(t, db, schema)
-	}
-	return
-}
-
-func getPsqlInfo(t *testing.T, container testcontainers.Container, outputDir string, query string) string {
-	i, err := container.Exec(
-		context.Background(),
-		[]string{"bash", "-c", "psql -U postgres -d " + *testDatabase + " -c '" + query + "' &> /testdata/output.out"},
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	output := readOutput(t, outputDir)
-
-	if i != 0 {
-		t.Logf("psql error. output: %s", output)
-	}
-	return output
-}
-
-func readOutput(t *testing.T, outputDir string) string {
-	outputFile := outputDir + "/output.out"
-	output, err := ioutil.ReadFile(outputFile)
-	if err != nil {
-		t.Errorf("error reading psql output: %v", err)
-	}
-	return string(output)
-}
-
-func getSchemas(t *testing.T, db *pgxpool.Pool) (out []string) {
-	row := db.QueryRow(
-		context.Background(),
-		"SELECT array_agg(nspname::TEXT order by nspname::TEXT) FROM pg_namespace",
-	)
-	err := row.Scan(&out)
-	if err != nil {
-		t.Errorf("could not discover schemas due to: %v", err)
-	}
-	return
-}
-
-func getTableInfosForSchema(t *testing.T, db *pgxpool.Pool, schema string) (out []tableInfo) {
-	row := db.QueryRow(
-		context.Background(),
-		"SELECT array_agg(relname::TEXT order by relname::TEXT) "+
-			"FROM pg_class "+
-			"WHERE relnamespace=$1::TEXT::regnamespace AND relkind='r'",
-		schema,
-	)
-	var tables []string
-	err := row.Scan(&tables)
-	if err != nil {
-		t.Errorf("could not get table info for schema \"%s\" due to: %v", schema, err)
-		return
-	}
-
-	out = make([]tableInfo, len(tables))
-	batch := pgx.Batch{}
-	for _, table := range tables {
-		batch.Queue(fmt.Sprintf(
-			"SELECT array_agg((tbl.*)::TEXT order by (tbl.*)::TEXT) from %s tbl",
-			pgx.Identifier{schema, table}.Sanitize(),
-		))
-	}
-	results := db.SendBatch(context.Background(), &batch)
-	defer results.Close()
-	for i, table := range tables {
-		out[i].name = table
-		err := results.QueryRow().Scan(&out[i].values)
-		if err != nil {
-			t.Errorf("error querying values from table %s: %v",
-				pgx.Identifier{schema, table}.Sanitize(), err)
-		}
-	}
-	return
 }
 
 // deep copy the metrics since we mutate them, and don't want to invalidate the tests
